refactor: update rpc service
