+++ packages/connection/src/common/constants.ts
@@ -1,1 +1,1 @@
-export const NOTREGISTERMETHOD = '$$NOTREGISTERMETHOD';
+export const METHOD_NOT_REGISTERED = '$$METHOD_NOT_REGISTERED';
