+++ packages/connection/src/common/index.ts
@@ -4,3 +4,4 @@ export * from './rpcProtocol';
 export * from './utils';
 export * from './ws-channel';
 export * from './connect';
+export * from './types';
