+++ packages/connection/src/common/rpc-service/index.ts
@@ -1,0 +1,2 @@
+export * from './stub';
+export * from './center';
