+++ packages/connection/src/common/rpc-service/stub.ts
@@ -1,0 +1,52 @@
+import { RPCServiceMethod, ServiceType } from '../types';
+import { getMethodName, getServiceMethods } from '../utils';
+
+import { RPCServiceCenter } from './center';
+
+export class RPCServiceStub {
+  constructor(private serviceName: string, private center: RPCServiceCenter, private type: ServiceType) {
+    if (this.type === ServiceType.Service) {
+      this.center.registerService(serviceName, this.type);
+    }
+  }
+
+  async ready() {
+    return this.center.when();
+  }
+
+  // 服务方
+  on(name: string, method: RPCServiceMethod) {
+    this.onRequest(name, method);
+  }
+
+  onRequestService(service: any) {
+    const methods = getServiceMethods(service);
+    for (const method of methods) {
+      this.onRequest(method, service[method].bind(service));
+    }
+  }
+
+  onRequest(name: string, method: RPCServiceMethod) {
+    this.center.onRequest(getMethodName(this.serviceName, name), method);
+  }
+
+  broadcast(name: string, ...args: any[]): Promise<any> {
+    return this.center.broadcast(getMethodName(this.serviceName, name), ...args);
+  }
+
+  getProxy = <T>() =>
+    new Proxy<T extends void ? RPCServiceStub : RPCServiceStub & T>(this as any, {
+      // 调用方
+      get: (target, prop: string) => {
+        if (!target[prop]) {
+          if (typeof prop === 'symbol') {
+            return Promise.resolve();
+          } else {
+            return (...args: any[]) => this.ready().then(() => this.broadcast(prop, ...args));
+          }
+        } else {
+          return target[prop];
+        }
+      },
+    });
+}
