+++ packages/connection/src/common/types.ts
@@ -1,3 +1,5 @@
+import { MessageConnection } from '@opensumi/vscode-jsonrpc';
+
 export interface ILogger {
   log(...args: any[]): void;
   warn(...args: any[]): void;
@@ -6,3 +8,20 @@ export interface ILogger {
 
 export type RPCServiceMethod = (...args: any[]) => any;
 export type IRPCServiceMap = Record<string, RPCServiceMethod>;
+
+export type ServiceProxy = any;
+
+export enum ServiceType {
+  Service,
+  Stub,
+}
+
+export interface IBench {
+  registerService: (service: string) => void;
+}
+
+export interface RPCMessageConnection extends MessageConnection {
+  uid?: string;
+  writer?: any;
+  reader?: any;
+}
